Protegendo rotas exclusivas para adminstradores

+++ backend/routes/mangas.ts
@@ -1,6 +1,8 @@
 import express from 'express';
+import { expressjwt, Request as JWTRequest } from "express-jwt"
 import { UploadedFile } from 'express-fileupload';
 import { MangaService } from '../services/mangaService';
+import { SECRET_KEY } from '../services/authService'
 
 const mangaService = new MangaService();
 
@@ -11,31 +13,35 @@ router.get('/', (req, res) => {
     res.status(200).json({ data: mangas});
 })
 
-router.post('/', (req, res) => {
-    const { data } = req.body
-    const { title, number, price }  = JSON.parse(data)
-    
-    if(title !== "" && !isNaN(number) && !isNaN(price)){
-        let coverPath = undefined 
-        if(req.files) {
-            try {
-                const cover = req.files['files.cover'] as UploadedFile;
-                coverPath = `/img/one_piece/upload/up${Date.now().toString()}-${cover.name}`
-                console.log(coverPath)
-                cover.mv(`${__dirname}/../../public/${coverPath}`)
-            } catch (err) {
-                res.status(500).json({ error: { message: err }});
-            }
-        }
+router.post('/', (req: JWTRequest, res) => {
+    if(req.auth?.role === "admin") {
+        const { data } = req.body
+        const { title, number, price }  = JSON.parse(data)
         
-        const addedManga = mangaService.create(title, Number(number), Number(price), coverPath)
-        if(addedManga) {
-            res.status(200).json({ data: addedManga })
+        if(title !== "" && !isNaN(number) && !isNaN(price)){
+            let coverPath = undefined 
+            if(req.files) {
+                try {
+                    const cover = req.files['files.cover'] as UploadedFile;
+                    coverPath = `/img/one_piece/upload/up${Date.now().toString()}-${cover.name}`
+                    console.log(coverPath)
+                    cover.mv(`${__dirname}/../../public/${coverPath}`)
+                } catch (err) {
+                    res.status(500).json({ error: { message: err }});
+                }
+            }
+            
+            const addedManga = mangaService.create(title, Number(number), Number(price), coverPath)
+            if(addedManga) {
+                res.status(200).json({ data: addedManga })
+            } else {
+                res.status(400).json({ error: { message: "Não foi possível adicionar o mangá!"}})
+            }
         } else {
-            res.status(400).json({ error: { message: "Não foi possível adicionar o mangá!"}})
+            res.status(400).json({ error: { message: "Informações inválidas!"}})
         }
     } else {
-        res.status(400).json({ error: { message: "Informações inválidas!"}})
+        res.status(403).json({ error: { message: "Forbidden" }})
     }
     
 });
@@ -52,42 +58,50 @@ router.get('/:id', (req, res) => {
     
 });
 
-router.put('/:id', (req, res) => {
-    const id = Number(req.params.id)
-    const { data } = req.body
-    const { title, number, price }  = JSON.parse(data)
-    
-    if(title !== "" && !isNaN(number) && !isNaN(price)){
-        let coverPath = undefined 
-        if(req.files) {
-            try {
-                const cover = req.files['files.cover'] as UploadedFile;
-                coverPath = `/img/one_piece/upload/up${Date.now().toString()}-${cover.name}`
-                console.log(coverPath)
-                cover.mv(`${__dirname}/../../public/${coverPath}`)
-            } catch (err) {
-                res.status(500).json({ error: { message: err }});
+router.put('/:id', expressjwt({ secret: SECRET_KEY, algorithms: ["HS256"] }), (req: JWTRequest, res) => {
+    if(req.auth?.role === "admin") {
+        const id = Number(req.params.id)
+        const { data } = req.body
+        const { title, number, price }  = JSON.parse(data)
+        
+        if(title !== "" && !isNaN(number) && !isNaN(price)){
+            let coverPath = undefined 
+            if(req.files) {
+                try {
+                    const cover = req.files['files.cover'] as UploadedFile;
+                    coverPath = `/img/one_piece/upload/up${Date.now().toString()}-${cover.name}`
+                    console.log(coverPath)
+                    cover.mv(`${__dirname}/../../public/${coverPath}`)
+                } catch (err) {
+                    res.status(500).json({ error: { message: err }});
+                }
             }
-        }
-        const manga = mangaService.update(id, title, Number(number), Number(price), coverPath)
-        if(manga) {
-            res.status(200).json({ data: manga });    
+            const manga = mangaService.update(id, title, Number(number), Number(price), coverPath)
+            if(manga) {
+                res.status(200).json({ data: manga });    
+            } else {
+                res.status(404).json({ error: { message: "Mangá não encontrado!"}})
+            }
         } else {
-            res.status(404).json({ error: { message: "Mangá não encontrado!"}})
+            res.status(400).json({ error: { message: "Informações inválidas!"}})
         }
     } else {
-        res.status(400).json({ error: { message: "Informações inválidas!"}})
+        res.status(403).json({ error: { message: "Forbidden" }})
     }
 
     
 });
 
-router.delete('/:id', (req, res) => {
-    const id = Number(req.params.id)
-    
-    if (mangaService.delete(id)) {
-        res.redirect('/mangas');
+router.delete('/:id', expressjwt({ secret: SECRET_KEY, algorithms: ["HS256"] }), (req: JWTRequest, res) => {
+    if(req.auth?.role === "admin") {
+        const id = Number(req.params.id)
+        
+        if (mangaService.delete(id)) {
+            res.redirect('/mangas');
+        } else {
+            res.status(404).json({ msg: "Mangá não encontrado!"});
+        }
     } else {
-        res.status(404).json({ msg: "Mangá não encontrado!"});
+        res.status(403).json({ error: { message: "Forbidden" }})
     }
 });
